format file
